leaves Balances view update

+++ IShop.Core/leaves_Days_track.cs
@@ -3,10 +3,10 @@
     public class leaves_Days_track
     {
         public int Id { get; set; }
-        public int Total_leaves_per_year { get; set; }
-        public int Requested_leaves { get; set; }
+        public decimal Total_leaves_per_year { get; set; }
+        public decimal Requested_leaves { get; set; }
         public string Type { get; set; }
-        public int Remaining_leaves { get; set; }
+        public decimal Remaining_leaves { get; set; }
         public string Username { get; set; }
     }
 }
