+++ IShop.Core/leaves_Types.cs
@@ -7,6 +7,6 @@ namespace IShop.Core
         public int Id { get; set; }
         public DateTime CreatedOn { get; set; }
         public string Type { get; set; }
-        public int Days { get; set; }
+        public Decimal Days { get; set; }
     }
 }
